requirement.txt
appnope==0.1.3
asttokens==2.0.5
backcall==0.2.0
backports.entry-points-selectable==1.1.1
click==8.1.2
coverage==6.3.2
decorator==5.1.1
distlib==0.3.4
executing==0.8.3
filelock==3.4.0
Flask==2.1.1
importlib-metadata==4.11.3
ipython==8.2.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mysql-connector-python==8.0.28
numpy==1.21.2
pandas==1.4.2
pandas-highcharts==0.5.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.0
prompt-toolkit==3.0.29
protobuf==3.20.1
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.11.2
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.35
stack-data==0.2.0
traitlets==5.1.1
virtualenv==20.10.0
wcwidth==0.2.5
Werkzeug==2.1.1
zipp==3.8.0
